fix(backend): 一次性对齐 m0044 家族外键/审计列模型 i64→i32（修实库 INT4 解码 DATABASE_ERROR 族）
- 根因: m0044 fix_fk_types() 运行期把 BIGINT REFERENCES <SERIAL主键白名单表> 静默改 INTEGER，
  实库 INT4 而模型仍 i64 → 这些表被 SELECT 解码时 mismatched types INT8 vs INT4 → 500(custom_order/sales_quotation 正被 extras 用例打到)
- 逐表模型 i64→i32: customer_color_price/custom_order(+sales_order_id)/after_sales/process_node/process_log(operator_id)/sales_quotation/seasonal_price_rule(product_category_id)
  及各 Create/Response/Update DTO 同步;保留非 m0044 来源或真边界的 i64(sales_user_id/created_by/approval_instance_id 等)
- service/handler 去为迁就错误 i64 而写的 as i64/冗余 as i32(quotation_approval/quotation_convert/quotation_ops crud/custom_order_state/aftersales handlers/customer_merge)
- m0044 迁移文本 BIGINT→INTEGER 自洽化(运行期行为不变);cargo fmt --check 0;只动这批 backend 文件(排除在途 bom)

+++ backend/src/handlers/custom_order_handler.rs
@@ -54,7 +54,7 @@ pub struct ListCustomOrdersQuery {
 #[allow(dead_code, reason = "反序列化输入字段")]
 #[derive(Debug, Deserialize)]
 pub struct AdvanceRequest {
-    pub operator_id: i64,
+    pub operator_id: i32,
     pub notes: Option<String>,
 }
