+++ backend/src/models/process_node.rs
@@ -18,7 +18,7 @@ pub struct Model {
     pub planned_end_date: Option<DateTime<Utc>>,
     pub actual_start_date: Option<DateTime<Utc>>,
     pub actual_end_date: Option<DateTime<Utc>>,
-    pub operator_id: Option<i64>,
+    pub operator_id: Option<i32>,
     pub notes: Option<String>,
     pub created_at: DateTime<Utc>,
     pub updated_at: DateTime<Utc>,
